Make source markers thinner and yellow
When the source density is high, the thick source markers obscure the
image.  Make the circles have a line width of 1 px.  To improve
visibility, change to colour from lime to yellow.  Yellow seems to be
the most visible colour, also when considering different kinds of colour
blindness and their incidence rates.